Do not find summaries during user interaction but find summaries during initialization

+++ __init__.py
@@ -18,44 +18,29 @@ class WebpageSummarizer(MycroftSkill):
                 c = conn.cursor()
                 c.execute('CREATE TABLE ? (url text, title text, summary text);', (self.table,))
                 conn.commit()
-                self.log.debug('Created database to store titles and summaries of URLs')
+                self.log.debug('Created database to store titles and summaries of URLs.')
+
+    def initialize(self):
+        self.summarize_webpages()
 
     @intent_file_handler('summarizer.webpage.intent')
     def handle_summarizer_webpage(self, message):
-        self.speak_dialog('summarizer.webpage')
-
         conn = sqlite3.connect(self.db)
         with conn:
             c = conn.cursor()
 
-            # Find summaries for all URLs in queue
-            c.execute('SELECT url FROM ? WHERE summary IS ?;', (self.table, None))
-            for row in c.fetchall():
-                url = row[0].strip()
-                title = summarized_web_text = ''
-                try:
-                    title, summarized_web_text = self.summarize_webpage(url)
-                    if title == '' or summarized_web_text == '':
-                        self.log.error('Couldn\'t fetch title or summary for the given URL: {}'.format(url))
-                        self.speak('Couldn not fetch title or summary for the given web page. I will continue to the next web page. Please check the logs for more details.')
-                        continue
-                    # Replace all | symbols before saving to the SQLite database
-                    title = title.replace('|', ',')
-                    summarized_web_text = summarized_web_text.replace('|', ',')
-                    c.execute('UPDATE ? SET title = ?, summary = ? WHERE url = ?;', (self.table, title, summarized_web_text, url))
-                    conn.commit()
-                except Exception as e:
-                    self.log.exception('Couldn\'t parse URL: {}.'.format(url))
-                    self.speak('Couldn not fetch title or summary for the given web page. I will continue to the next web page. Please check the logs for more details.')
-                    continue
-
             # Read out all the summaries in the queue and clear them out
             c.execute('SELECT * FROM ? WHERE summary IS NOT ?;', (self.table, None))
-            for row in c.fetchall():
-                self.speak('Web page title is {}'.format(row[1]))
-                self.speak('And the summary is as follows. {}'.format(row[2]))
-                c.execute('DELETE from ? where url = ?;', (self.table, row[0]))
-                conn.commit()
+            data_to_speak = c.fetchall()
+            if len(data_to_speak) > 0:
+                for row in data_to_speak:
+                    self.speak_dialog('summarizer.webpage')
+                    self.speak('Web page title is {}'.format(row[1]))
+                    self.speak('And the summary is as follows. {}'.format(row[2]))
+                    c.execute('DELETE from ? where url = ?;', (self.table, row[0]))
+                    conn.commit()
+            else:
+                self.speak('No more summaries available.')
 
     def summarize_webpage(self, url):
         """
@@ -70,13 +55,41 @@ class WebpageSummarizer(MycroftSkill):
             website_text = ' '.join(map(lambda p: p.text, page.find_all('p')))
             title = page.title.text.strip()
             summarized_text = summarize(website_text).strip()
-            self.log.debug('\n\nTitle: {}\nSummary: {}\n\n'.format(title, summarized_text))
+            self.log.debug('\n\nURL: {}\nTitle: {}\nSummary: {}\n\n'.format(url, title, summarized_text))
         except Exception as e:
             self.log.exception('Couldn\'t parse URL: {}.'.format(url))
         finally:
             self.browser.close()
 
         return title, summarized_text
+
+    def summarize_webpages(self):
+        conn = sqlite3.connect(self.db)
+        with conn:
+            c = conn.cursor()
+
+            # Find summaries for all URLs in queue
+            c.execute('SELECT url FROM ? WHERE summary IS ?;', (self.table, None))
+            data_to_summarize = c.fetchall()
+            if len(data_to_summarize > 0):
+                for row in data_to_summarize:
+                    url = row[0].strip()
+                    title = summarized_web_text = ''
+                    try:
+                        title, summarized_web_text = self.summarize_webpage(url)
+                        if title == '' or summarized_web_text == '':
+                            self.log.error('Couldn\'t fetch title or summary for the given URL: {}'.format(url))
+                            continue
+                        # Replace all '|' symbols before saving to the SQLite database
+                        title = title.replace('|', ',')
+                        summarized_web_text = summarized_web_text.replace('|', ',')
+                        c.execute('UPDATE ? SET title = ?, summary = ? WHERE url = ?;', (self.table, title, summarized_web_text, url))
+                        conn.commit()
+                    except Exception as e:
+                        self.log.exception('Couldn\'t parse URL: {}.'.format(url))
+                        continue
+            else:
+                self.log.debug('No pending URLs to summarize.')
 
 
 def create_skill():
